Setting default domain for V3 ListUser Identity-command
Based on keystone #1498556 is safe to assume the default domain
can be specified as default.
Setting the default domain to 'Default' for ListUser
openstack commands will fix the rather annoying issue
for end users to specify a domain even when just using
the defaults in keystone. That is particular important
when enabling multi domain drivers in Keystone and LDAP.
Lastly the Domain Id colume is added to any standard
ListUser output

Closes-Bug: #1521420
Change-Id: Ib174b7461f6f2a23b2c6f14c58756bbd816fc6bf